ruff now run on actual code && pytest tests
