+++ .github/workflows/python-app.yml
@@ -23,37 +23,41 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - name: Set up Python 3.12
-      uses: actions/setup-python@v3
-      with:
-        python-version: "3.12"
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install ruff
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-    - name: Lint with ruff
-      run: |
-        ruff check
-    - name: Format with ruff
-      run: |
-        ruff format --check --diff
+      - uses: actions/checkout@v4
+      - name: Set up Python 3.12
+        uses: actions/setup-python@v3
+        with:
+          python-version: "3.12"
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install ruff
+          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+      - name: Lint with ruff
+        run: |
+          ruff check .
+        working-directory: ./backend
+  #    - name: Format with ruff
+  #      run: |
+  #        ruff format --check --diff .
+  #      working-directory: ./backend
 
-  # unit_tests:
+  unit_tests:
 
-  #   runs-on: ubuntu-latest
+     runs-on: ubuntu-latest
 
-  #   steps:
-  #   - uses: actions/checkout@v4
-  #   - name: Set up Python 3.12
-  #     uses: actions/setup-python@v3
-  #     with:
-  #       python-version: "3.12"
-  #   - name: Install dependencies
-  #     run: |
-  #       python -m pip install --upgrade pip
-  #       pip install pytest
-  #       if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-  #   - name: Test with pytest
-  #     run: |
-  #       pytest
+     steps:
+       - uses: actions/checkout@v4
+       - name: Set up Python 3.12
+         uses: actions/setup-python@v3
+         with:
+           python-version: "3.12"
+       - name: Install dependencies
+         run: |
+           python -m pip install --upgrade pip
+           pip install pytest
+           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+       - name: Test with pytest
+         run: |
+           pytest
+         working-directory: ./backend
