+++ backend/unit_tests/test_book_create.py
@@ -1,9 +1,5 @@
 import pytest
-import sys
 from fastapi.testclient import TestClient
-
-from .tools import *
-
 
 @pytest.mark.parametrize(
     "data",
