+++ backend/unit_tests/test_book_delete.py
@@ -1,8 +1,4 @@
-import sys
 from fastapi.testclient import TestClient
-
-from .tools import *
-
 
 def test_delete_unknown_book(client: TestClient) -> None:
     response = client.delete("/api/v1/books/1")
