+++ backend/unit_tests/test_book_read_all.py
@@ -1,8 +1,4 @@
-import sys
 from fastapi.testclient import TestClient
-
-from .tools import *
-
 
 def test_read_all_book_without_book(client: TestClient) -> None:
     response = client.get("/api/v1/books")
